# refactor normalizeArray, improve performance