--- a/lib/path.js
+++ b/lib/path.js
@@ -30,34 +30,45 @@ var util = require('util');
 // relative and absolute paths)
 function normalizeArray(parts, allowAboveRoot) {
   var nonEmptyParts = [];
+  var nonBack = true;
   for (var i = 0; i < parts.length; i++) {
     var p = parts[i];
     if (p && p !== '.') {
       nonEmptyParts.push(p);
     }
+    if (p === '..') {
+      nonBack = false;
+    }
   }
+
   parts = nonEmptyParts;
 
-  // if the path tries to go above the root, `up` ends up > 0
+  // if the path does not contain ..
+  if (nonBack) {
+    return parts;
+  }
+
+  // if the path tries to go ab ove the root, `up` ends up > 0
   var up = 0;
+  var res = [];
   for (var i = parts.length - 1; i >= 0; i--) {
     if (parts[i] === '..') {
-      parts.splice(i, 1);
       up++;
     } else if (up) {
-      parts.splice(i, 1);
       up--;
+    } else {
+      res.push(parts[i]);
     }
   }
 
   // if the path is allowed to go above the root, restore leading ..s
   if (allowAboveRoot) {
     for (; up--; up) {
-      parts.unshift('..');
+      res.push('..');
     }
   }
 
-  return parts;
+  return res.reverse();
 }
 
 if (isWindows) {
@@ -335,7 +346,8 @@ if (isWindows) {
     // handle relative paths to be safe (might happen when process.cwd() fails)
 
     // Normalize the path
-    resolvedPath = normalizeArray(resolvedPath.split('/'), !resolvedAbsolute).join('/');
+    resolvedPath = normalizeArray(resolvedPath.split('/'),
+                                  !resolvedAbsolute).join('/');
 
     return ((resolvedAbsolute ? '/' : '') + resolvedPath) || '.';
   };
@@ -344,7 +356,7 @@ if (isWindows) {
   // posix version
   exports.normalize = function(path) {
     var isAbsolute = exports.isAbsolute(path),
-        trailingSlash = path[path.length - 1] === '/',
+        trailingSlash = path[path.length - 1] === '/';
 
     // normalize the path
     path = normalizeArray(path.split('/'), !isAbsolute).join('/');