@dead-horse 2015-08-18
it is already fixed by https://github.com/joyent/node/pull/8724. :)